norminette done
