+++ extraordanary_cases.txt
@@ -35,3 +35,10 @@ kill process: kill -9 pid(pid number sofia)
 
 
 echo - "" "" "hello"
+
+
+<< ola | ls | wc
+
+control_d mario 
+
+protecao mallocs
