+++ srcs/executer/heredoc.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_atoi.c                                          :+:      :+:    :+:   */
+/*   heredoc.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: svalente <svalente@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/08 16:41:17 by svalente          #+#    #+#             */
-/*   Updated: 2023/11/13 12:31:08 by svalente         ###   ########.fr       */
+/*   Updated: 2023/11/13 21:09:26 by svalente         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,11 +23,10 @@ char	*expansion_heredoc(char *s)
 		if (s[j] && s[j] == '$')
 		{
 			tmp = check_expansion(s, &j);
-	      	free(s);
+			free(s);
 			s = tmp;
 		}
 	}
-	printf("herdoc expand: %s\n", s);
 	return (s);
 }
