+++ srcs/print_functions.c
@@ -31,3 +31,37 @@ void	print_list(t_cmd *lst)
 	}
 	printf("---------------------\n");
 }
+
+void	print_matrix(char **m)
+{
+	int	i;
+
+	i = 0;
+	if (!m)
+		return ;
+	while (m[i])
+	{
+		printf("[%s]\n", m[i]);
+		i++;
+	}
+	printf("\n");
+}
+
+void	print_redir(t_redir *lst)
+{
+	t_redir *tmp = lst;
+
+	if (!lst)
+	{
+		printf("Empty list for redirections\n");	
+		return ;
+	}
+	while (tmp)
+	{
+		// printf("Address of red: %p\n", tmp);
+		// printf("Address of file: %p\n", tmp->file);
+		printf("redir: %d\n", tmp->redir);
+		printf("file: %s\n", tmp->file);
+		tmp = tmp->next;
+	}
+}
